chore: move form-publisher component
