+++ src/components/form-publisher/configured-form-publisher.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { reduxForm } from 'redux-form';
 
 import FormPublisher from './form-publisher.component';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'publisher',
