+++ src/components/form-publisher/field-publisher/field-publisher.component.tsx
@@ -2,7 +2,7 @@ import React, { memo, FC } from 'react';
 import { compose } from 'redux';
 import type { WrappedFieldProps } from 'redux-form';
 
-import Translation from '../../../../components/translation';
+import Translation from '../../translation';
 
 interface ExternalProps {
   label: string;
