+++ src/components/form-publisher/search-publisher/search-publisher-resolver.js
@@ -1,5 +1,5 @@
 import { resolve } from 'react-resolver';
-import { memoizedGetOrganizationList } from '../../../../services/api/organization-api/host';
+import { memoizedGetOrganizationList } from '../../../services/api/organization-api/host';
 
 const mapProps = {
   suggestions: memoizedGetOrganizationList
