+++ src/components/form-publisher/search-publisher/search-publisher.component.js
